pyproject.toml
[tool.poetry]
name = "dz1"
version = "0.1.0"
description = ""
authors = ["hhssome <artempavel-terehove@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.89.1"
pydantic = {extras = ["dotenv"], version = "^1.10.4"}
alembic = "^1.9.1"
uvicorn = "^0.20.0"
asyncpg = "^0.27.0"
sqlalchemy2-stubs = "^0.0.2a31"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
